% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narxdataprep.R
\name{CoFESSNARXdata}
\alias{CoFESSNARXdata}
\title{CoFESSNARXdata}
\usage{
CoFESSNARXdata(yy, xx, ydelay, xdelay, delay = 1, s = NULL, date = T)
}
\arguments{
\item{yy}{numeric vectors containing target data}

\item{xx}{numeric vectors containing feature data}

\item{ydelay}{integer representing the desired AR order for y}

\item{xdelay}{integer representing the desired AR order for x}

\item{delay}{integers representing n, the size of the prediction window into the future}

\item{s}{integers representing the seasonality desired.  Can be expressed as a
list for P>1.}

\item{date}{boolean value to indicate if you data includes dates}
}
\value{
returns the target combined with the seasonallly lagged target variable,and the original
and lagged exegenous variables in one database.
}
\description{
\code{CoFESSNARXdata} returns the lead/lag series necessary to run the CoFESSNARX
}
\details{
This function transforms the target and exegenous series variables for the
 CoFESSNARX model.  Variables series, and the number of desired lags for each
 group respectively and then returns the requested lag series arlong with the
 seasonally lagged verson of teh original series for each in a single  data set.
}
